FM-579 Add tenant prefix where dynamic_fork is used
We are using dynamic fork encapsulated in a DYNAMIC_FORK workflow that
always receives tasks to be dynamically executed via dynamic_tasks
object.

This commit adds tenant prefix to task names inside dynamic_tasks (this
convention needs to be upheld).

Also, there is a sanity check that dynamic_tasks contain a single type
of task. This sanity chack is part of DYNAMIC_FORK workflow and uses
expectedName paramter. This commit also adds tenant prefix to input
parameter named expectedName (this convention also needs to be upheld).

+ Add unit test run to Dockerfil build for wf proxy

Signed-off-by: Maros Marsalek <mmarsalek@frinx.io>